+++ engine/src/flutter/shell/platform/android/io/flutter/view/accessibility/configurator/ClassNameRoleConfigurator.java
@@ -1,0 +1,22 @@
+// Copyright 2013 The Flutter Authors. All rights reserved.
+// Use of this source code is governed by a BSD-style license that can be
+// found in the LICENSE file.
+
+package io.flutter.view.accessibility.configurator;
+
+import android.view.accessibility.AccessibilityNodeInfo;
+import io.flutter.view.AccessibilityBridge;
+
+/** Configurator that simply sets the class name of the accessibility node. */
+public class ClassNameRoleConfigurator implements AccessibilityNodeConfigurator {
+  private final String className;
+
+  public ClassNameRoleConfigurator(String className) {
+    this.className = className;
+  }
+
+  @Override
+  public void configure(AccessibilityNodeInfo result, AccessibilityBridge.SemanticsNode node) {
+    result.setClassName(className);
+  }
+}
